refactor(helper): extract header encoding utilities from HttpConnection into HttpHeaderEncoding
